move filter from HaeinsaDelete to HaeinsaRowTransaction
