+++ src/main/java/kr/co/vcnc/haeinsa/HaeinsaRowTransaction.java
@@ -16,7 +16,10 @@
 package kr.co.vcnc.haeinsa;
 
 import java.util.List;
+import java.util.Map;
+import java.util.NavigableSet;
 
+import com.google.common.annotations.VisibleForTesting;
 import kr.co.vcnc.haeinsa.thrift.generated.TRowLock;
 
 import com.google.common.collect.Lists;
@@ -94,7 +97,7 @@ class HaeinsaRowTransaction {
         }
 
         HaeinsaDelete lastDelete = (HaeinsaDelete) mutations.remove(mutations.size() - 1);
-        HaeinsaDelete.FilterResult filterResult = lastDelete.filter(put);
+        FilterResult filterResult = filter(lastDelete, put);
         mutations.get(mutations.size() - 1).add(filterResult.getRemained());
         mutations.add(lastDelete);
         if (!filterResult.getDeleted().isEmpty()) {
@@ -116,11 +119,28 @@ class HaeinsaRowTransaction {
 
         // prevDelete, lastPut + newDelete => (prevDelete + newDelete), remainedPut
         HaeinsaPut lastPut = (HaeinsaPut) mutations.remove(mutations.size() - 1);
-        HaeinsaDelete.FilterResult filterResult = delete.filter(lastPut);
+        FilterResult filterResult = filter(delete, lastPut);
         mutations.get(mutations.size() - 1).add(delete);
         if (!filterResult.getRemained().isEmpty()) {
             mutations.add(filterResult.getRemained());
         }
+    }
+
+    @VisibleForTesting
+    static FilterResult filter(HaeinsaDelete delete, HaeinsaPut put) {
+        HaeinsaDeleteTracker deleteTracker = new HaeinsaDeleteTracker();
+        deleteTracker.add(delete, 0);
+        FilterResult filterResult = new FilterResult(delete.getRow());
+        for (Map.Entry<byte[], NavigableSet<HaeinsaKeyValue>> entry : put.getFamilyMap().entrySet()) {
+            for (HaeinsaKeyValue keyValue : entry.getValue()) {
+                if (deleteTracker.isDeleted(keyValue, 1)) {
+                    filterResult.getDeleted().add(keyValue.getFamily(), keyValue.getQualifier(), keyValue.getValue());
+                } else {
+                    filterResult.getRemained().add(keyValue.getFamily(), keyValue.getQualifier(), keyValue.getValue());
+                }
+            }
+        }
+        return filterResult;
     }
 
     public HaeinsaTableTransaction getTableTransaction() {
@@ -139,5 +159,24 @@ class HaeinsaRowTransaction {
             result.add(mutation.getScanner(mutations.size() - i));
         }
         return result;
+    }
+
+    @VisibleForTesting
+    static final class FilterResult {
+        private final HaeinsaPut remained;
+        private final HaeinsaPut deleted;
+
+        private FilterResult(byte[] row) {
+            this.remained = new HaeinsaPut(row);
+            this.deleted = new HaeinsaPut(row);
+        }
+
+        public HaeinsaPut getRemained() {
+            return remained;
+        }
+
+        public HaeinsaPut getDeleted() {
+            return deleted;
+        }
     }
 }
